+++ ci/get-version.sh
@@ -5,6 +5,6 @@ set -eufo pipefail
 cd $(dirname "$0")
 cd ..
 
-version=$(grep -E -m 1 "^version" comsrv/Cargo.toml | cut -d "=" -f 2 | sed 's/[" ]//g')
+version=$(grep -m 1 -oE '## \[.*?\]' CHANGELOG.md | sed -e 's/[# \[]//g' -e 's/\]//g')
 
 echo $version