+++ python/comsrv/test/test_import.py
@@ -1,4 +1,4 @@
-def test_import():
+def test_import() -> None:
     from comsrv import ByteStreamInstrument  # noqa: F401
     from comsrv import CanBus  # noqa: F401